Serve Whitehall organisation logos from Asset Manager
This routes requests for Whitehall organisations logos to Asset Manager
instead of to Whitehall. See
https://github.com/alphagov/asset-manager/milestone/9 for more info.

The change to common.yaml results in a new `location` block being added
to the assets-origin nginx config to proxy the request to static.

The change to static_extra_nginx_config.conf.erb adds a `location` block
to proxy the request to asset-manager. The additional configuration is a
duplication of the configuration required to proxy requests for `/media`
to static in the same file. I plan to DRY this duplication up in a
subsequent commit.

Requests for organisation logos were previously being proxied from
assets-origin to whitehall-frontend by the `/government/uploads`
location configured in hierdata/common.yaml.

Note that Nginx matches longer prefix matches in preference to shorter
matches[1], which is why requests for organisations logos will be
handled by our new location block instead of the `/government/uploads/`
block that was already in place.

I've tested the effect of this change on my dev-vm and am confident that
it's working as expected. I saw the following differences in response
headers but I don't think we need to worry about any of them:

* Cache-Control
  * `max-age=14400, public` from Whitehall
  * `private` from Asset Manager
  * This is expected because I'm running Asset Manager in development
  * mode. It'll add the correct `Cache-Control` headers in production.

* Last-Modified
  * `Wed, 25 Oct 2017 15:01:08 GMT` from Whitehall
  * `Wed, 25 Oct 2017 15:00:34 GMT` from Asset Manager
  * This is expected because this was a new image uploaded to Whitehall
  * which was then sent to Asset Manager so it'll have been written to
  * the two different places on disk at different times.

* ETag
  * `"59f0a734-13767"` from Whitehall
  * `"59f0a712-13767"` from Asset Manager
  * This is expected because this was a new image uploaded to Whitehall
  * which was then sent to Asset Manager. Part of this etag is based on
  * the Last Modified time which, as explained above, is also different.

* Strict-Transport-Security
  * Not present from Whitehall
  * `max-age=31536000` from Asset Manager
  * This is added by the `location` block in the static nginx
  * configuration.
  * It seems fine to add this header given that we want to encourage
  * browsers to always access this site over HTTPS.

[1]:
https://www.digitalocean.com/community/tutorials/understanding-nginx-server-and-location-block-selection-algorithms